Split rpcHandler out of cmd/mesh/main.go
